// lib/model/attemptSetupAttempt.ts
/**
 * Fiber Swagger Example API
 * This is a sample server server.
 *
 * The version of the OpenAPI document: 2.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AttemptStatus } from './attemptStatus';
import { AttemptUsage } from './attemptUsage';
import { DatabaseDeletedAt } from './databaseDeletedAt';

export class AttemptSetupAttempt {
    'attachToSelf'?: boolean;
    'createdAt'?: string;
    'customer'?: string;
    'deletedAt'?: DatabaseDeletedAt;
    'flowDirections'?: Array<string>;
    'id'?: string;
    'livemode'?: boolean;
    'metadata'?: object;
    'object'?: string;
    'paymentMethod'?: string;
    'paymentMethodDetails'?: string;
    'setupError'?: string;
    'setupIntent'?: string;
    'status'?: AttemptStatus;
    'updatedAt'?: string;
    'usage'?: AttemptUsage;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "attachToSelf",
            "baseName": "attach_to_self",
            "type": "boolean"
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "string"
        },
        {
            "name": "customer",
            "baseName": "customer",
            "type": "string"
        },
        {
            "name": "deletedAt",
            "baseName": "deleted_at",
            "type": "DatabaseDeletedAt"
        },
        {
            "name": "flowDirections",
            "baseName": "flow_directions",
            "type": "Array<string>"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "livemode",
            "baseName": "livemode",
            "type": "boolean"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "object"
        },
        {
            "name": "object",
            "baseName": "object",
            "type": "string"
        },
        {
            "name": "paymentMethod",
            "baseName": "payment_method",
            "type": "string"
        },
        {
            "name": "paymentMethodDetails",
            "baseName": "payment_method_details",
            "type": "string"
        },
        {
            "name": "setupError",
            "baseName": "setup_error",
            "type": "string"
        },
        {
            "name": "setupIntent",
            "baseName": "setup_intent",
            "type": "string"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "AttemptStatus"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "string"
        },
        {
            "name": "usage",
            "baseName": "usage",
            "type": "AttemptUsage"
        }    ];

    static getAttributeTypeMap() {
        return AttemptSetupAttempt.attributeTypeMap;
    }
}

export namespace AttemptSetupAttempt {
}