// lib/model/bankBankAccount.ts
export class BankBankAccount {
    'account'?: AccountAccount;
    'accountHolderName'?: string;
    'accountHolderType'?: BankAccountHolderType;
    'accountId'?: UuidNullUUID;
    'accountType'?: BankAccountType;
    'availablePayoutMethods'?: Array<string>;
    'bankName'?: string;
    'country'?: string;
    'createdAt'?: string;
    'currency'?: string;
    'customer'?: CustomerCustomer;
    'customerId'?: UuidNullUUID;
    'defaultForCurrency'?: boolean;
    'deletedAt'?: DatabaseDeletedAt;
    'fingerprint'?: string;
    'id'?: string;
    'last4'?: string;
    'livemode'?: boolean;
    'metadata'?: object;
    'object'?: string;
    'routingNumber'?: string;
    'status'?: BankStatus;
    'updatedAt'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "account",
            "baseName": "account",
            "type": "AccountAccount"
        },
        {
            "name": "accountHolderName",
            "baseName": "account_holder_name",
            "type": "string"
        },
        {
            "name": "accountHolderType",
            "baseName": "account_holder_type",
            "type": "BankAccountHolderType"
        },
        {
            "name": "accountId",
            "baseName": "account_id",
            "type": "UuidNullUUID"
        },
        {
            "name": "accountType",
            "baseName": "account_type",
            "type": "BankAccountType"
        },
        {
            "name": "availablePayoutMethods",
            "baseName": "available_payout_methods",
            "type": "Array<string>"
        },
        {
            "name": "bankName",
            "baseName": "bank_name",
            "type": "string"
        },
        {
            "name": "country",
            "baseName": "country",
            "type": "string"
        },
        {
            "name": "createdAt",
            "baseName": "created_at",
            "type": "string"
        },
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string"
        },
        {
            "name": "customer",
            "baseName": "customer",
            "type": "CustomerCustomer"
        },
        {
            "name": "customerId",
            "baseName": "customer_id",
            "type": "UuidNullUUID"
        },
        {
            "name": "defaultForCurrency",
            "baseName": "default_for_currency",
            "type": "boolean"
        },
        {
            "name": "deletedAt",
            "baseName": "deleted_at",
            "type": "DatabaseDeletedAt"
        },
        {
            "name": "fingerprint",
            "baseName": "fingerprint",
            "type": "string"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "last4",
            "baseName": "last4",
            "type": "string"
        },
        {
            "name": "livemode",
            "baseName": "livemode",
            "type": "boolean"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "object"
        },
        {
            "name": "object",
            "baseName": "object",
            "type": "string"
        },
        {
            "name": "routingNumber",
            "baseName": "routing_number",
            "type": "string"
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "BankStatus"
        },
        {
            "name": "updatedAt",
            "baseName": "updated_at",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return BankBankAccount.attributeTypeMap;
    }
}

export namespace BankBankAccount {
}